Finally fixed long-standing annoying bug triggered by route change detection.  Bug was triggered when:
1. A previously routed network is detected as having been assigned to a new
   tunnel, and
2. The losing tunnel's inner address comes entirely from the routed network it
   it is losing.

This particular change requires a "rebase" operation, in which the losing
tunnel is brought down, brought back up, and then reassigned all of the
networks that it previously had, minus the network that is migrating away from
it.

However, due to a typo, while the rebasing logic thought it was adding all the
old undisturbed networks back to the tunnel, it in fact was adding the very
network it was supposed to loose in the first place, and not just once, but
over and over again.

(Use of wrong variable, "route" when it should have used "other").